1:1 관리자 문의게시판 유효성 검사 완료

+++ src/main/java/com/heydoctor/app/service/adminpage/QnaPageServiceImpl.java
@@ -30,7 +30,7 @@ public class QnaPageServiceImpl implements QnaPageService{
     }
 
     @Override
-    public void delete(int qnaId) {
+    public void delete(Long qnaId) {
         qnaDAO.deleteQna(qnaId);
     }
 }
